Bring layout initialization up to AppInit
Cleans it up a little bit
